Conflicts:
	Examples/Cxx/ConvertToQImage.cxx
	Examples/Cxx/FixJAIBugJPEGLS.cxx
	Source/DataStructureAndEncodingDefinition/gdcmExplicitDataElement.txx
	Source/MediaStorageAndFileFormat/gdcmJPEG2000Codec.cxx
	Source/MediaStorageAndFileFormat/gdcmJPEGCodec.cxx
	Source/MediaStorageAndFileFormat/gdcmRLECodec.cxx
	Source/MediaStorageAndFileFormat/gdcmStreamImageReader.cxx
	Testing/Source/Data/CMakeLists.txt
	Testing/Source/MediaStorageAndFileFormat/Cxx/TestIPPSorter3.cxx
	Utilities/VTK/Applications/CMakeLists.txt
	Utilities/VTK/Testing/Cxx/CMakeLists.txt
	Utilities/VTK/Testing/Java/CMakeLists.txt
	Wrapping/CMakeLists.txt
	Wrapping/Csharp/CMakeLists.txt
	Wrapping/Perl/CMakeLists.txt

+++ Applications/Cxx/gdcmdump.cxx
@@ -139,7 +139,7 @@ void printbinary(std::istream &is, PDFElement const & pdfel )
   uint32_t dummy = pdfel.getdummy();
   assert( dummy == 0 );
   uint32_t offset = pdfel.getoffset();
-  uint32_t pos = offset + is.tellg() - 4;
+  uint32_t pos = (uint32_t)(offset + is.tellg() - 4);
   printvalue(is, type, numels, pos);
 }
