+++ Examples/Cxx/GenSeqs.cxx
@@ -65,7 +65,7 @@ int main(int argc, char *argv[])
 
   const char owner_str[] = "GDCM CONFORMANCE TESTS";
   gdcm::DataElement owner( gdcm::Tag(0x4d4d, 0x10) );
-  owner.SetByteValue(owner_str, strlen(owner_str));
+  owner.SetByteValue(owner_str, (uint32_t)strlen(owner_str));
   owner.SetVR( gdcm::VR::LO );
 
   for(unsigned int idx = 0; idx < 10/* nitems*/; ++idx)
