+++ Source/DataDictionary/gdcmGroupDict.h
@@ -69,7 +69,7 @@ inline std::ostream& operator<<(std::ostream& _os, const GroupDict &_val)
   for(size_t i=0; i<size; ++i)
     {
     _os << std::hex << std::setw(4) << std::setfill( '0' ) << i << ","
-      << _val.GetAbbreviation(i) << "," << _val.GetName(i) << "\n";
+      << _val.GetAbbreviation((uint16_t)i) << "," << _val.GetName((uint16_t)i) << "\n";
     }
   return _os;
 }
