+++ Testing/Source/DataStructureAndEncodingDefinition/Cxx/TestVM.cxx
@@ -19,19 +19,19 @@ int TestVM(int, char *[])
   //gdcm::VM::VMType vm = gdcm::LengthToVM<1>::TVM;
 
   const char str1[] = " 1\\2";
-  unsigned int count1 = gdcm::VM::GetNumberOfElementsFromArray(str1, strlen(str1) );
+  unsigned int count1 = gdcm::VM::GetNumberOfElementsFromArray(str1, (unsigned int)strlen(str1) );
   if( count1 != 2 ) return 1;
 
   const char str2[] = "  1\\2\\3";
-  unsigned int count2 = gdcm::VM::GetNumberOfElementsFromArray(str2, strlen(str2) );
+  unsigned int count2 = gdcm::VM::GetNumberOfElementsFromArray(str2, (unsigned int)strlen(str2) );
   if( count2 != 3 ) return 1;
 
   const char str3[] = "   1";
-  unsigned int count3 = gdcm::VM::GetNumberOfElementsFromArray(str3, strlen(str3) );
+  unsigned int count3 = gdcm::VM::GetNumberOfElementsFromArray(str3, (unsigned int)strlen(str3) );
   if( count3 != 1 ) return 1;
 
   const char str4[] = "";
-  unsigned int count4 = gdcm::VM::GetNumberOfElementsFromArray(str4, strlen(str4) );
+  unsigned int count4 = gdcm::VM::GetNumberOfElementsFromArray(str4, (unsigned int)strlen(str4) );
   if( count4 != 0 ) return 1;
 
   const char *str5 = 0;
@@ -39,15 +39,15 @@ int TestVM(int, char *[])
   if( count5 != 0 ) return 1;
 
   const char str6[] = "   1";
-  unsigned int count6 = gdcm::VM::GetNumberOfElementsFromArray(str6, strlen(str6) );
+  unsigned int count6 = gdcm::VM::GetNumberOfElementsFromArray(str6, (unsigned int)strlen(str6) );
   if( count6 != 1 ) return 1;
 
   const char str7[] = "   1 \\   2  ";
-  unsigned int count7 = gdcm::VM::GetNumberOfElementsFromArray(str7, strlen(str7) );
+  unsigned int count7 = gdcm::VM::GetNumberOfElementsFromArray(str7, (unsigned int)strlen(str7) );
   if( count7 != 2 ) return 1;
 
   const char str8[] = "   ";
-  unsigned int count8 = gdcm::VM::GetNumberOfElementsFromArray(str8, strlen(str8) );
+  unsigned int count8 = gdcm::VM::GetNumberOfElementsFromArray(str8, (unsigned int)strlen(str8) );
   if( count8 != 0 )
     {
     std::cerr << "count8 failed" << std::endl;
@@ -55,11 +55,11 @@ int TestVM(int, char *[])
     }
 
   const char str9[] = "     \\      ";
-  unsigned int count9 = gdcm::VM::GetNumberOfElementsFromArray(str9, strlen(str9) );
+  unsigned int count9 = gdcm::VM::GetNumberOfElementsFromArray(str9, (unsigned int)strlen(str9) );
   if( count9 != 0 ) return 1;
 
   const char str10[] = "   3  \\      ";
-  unsigned int count10 = gdcm::VM::GetNumberOfElementsFromArray(str10, strlen(str10) );
+  unsigned int count10 = gdcm::VM::GetNumberOfElementsFromArray(str10, (unsigned int)strlen(str10) );
   if( count10 != 1 ) return 1;
 
   if( gdcm::VM::VM1 & gdcm::VM::VM2 ) return 1;
