+++ Utilities/VTK/Testing/Cxx/TestvtkGDCMImageReader.cxx
@@ -37,7 +37,7 @@ int TestvtkGDCMImageRead(const char *filename, bool verbose)
     gdcm::Directory d;
     d.Load( filename );
     gdcm::Directory::FilenamesType l = d.GetFilenames();
-    const unsigned int nfiles = l.size();
+    const size_t nfiles = l.size();
     vtkStringArray *sarray = vtkStringArray::New();
     for(unsigned int i = 0; i < nfiles; ++i)
       {
